+++ app/assets/javascripts/discourse/app/templates/application.hbs
@@ -1,3 +1,5 @@
+<DStyles />
+
 <DiscourseRoot>
   <a href="#main-container" id="skip-link">{{i18n "skip_to_main_content"}}</a>
   <DDocument />
